@@ -20,8 +20,8 @@ if (!defined('IN_PHPBB'))
 }
 
 /**
-* @todo [smilies] check regular expressions for special char replacements (stored specialchared in db)
-*/
+  * @todo {smilies} check regular expressions for special char replacements (stored specialchared in db)
+  */
 class acp_icons
 {
 	var $u_action;
