@@ -20,8 +20,8 @@ if (!defined('IN_PHPBB'))
 }
 
 /**
-* @todo [words] check regular expressions for special char replacements (stored specialchared in db)
-*/
+  * @todo {words} check regular expressions for special char replacements (stored specialchared in db)
+  */
 class acp_words
 {
 	var $u_action;
