@@ -208,16 +208,16 @@ class postgres_extractor extends base_extractor
 		}
 
 		// Generate constraint clauses for CHECK constraints
-		$sql_checks = "SELECT conname as index_name, consrc
-			FROM pg_constraint, pg_class bc
-			WHERE conrelid = bc.oid
+		$sql_checks = "SELECT pc.conname AS index_name, pg_get_constraintdef(pc.oid)
+			FROM pg_constraint pc, pg_class bc
+			WHERE pc.conrelid = bc.oid
 				AND bc.relname = '" . $this->db->sql_escape($table_name) . "'
 				AND NOT EXISTS (
 					SELECT *
-						FROM pg_constraint as c, pg_inherits as i
-						WHERE i.inhrelid = pg_constraint.conrelid
-							AND c.conname = pg_constraint.conname
-							AND c.consrc = pg_constraint.consrc
+					FROM pg_constraint AS c, pg_inherits AS i
+						WHERE i.inhrelid = pc.conrelid
+							AND c.conname = pc.conname
+							AND pg_get_constraintdef(c.oid) = pg_get_constraintdef(pc.oid)
 							AND c.conrelid = i.inhparent
 				)";
 		$result = $this->db->sql_query($sql_checks);
